+++ api/src/main/scala/dx/api/DxIoParameter.scala
@@ -79,6 +79,7 @@ final case class IOParameterValueString(value: String) extends IOParameterValue
 final case class IOParameterValueNumber(value: BigDecimal) extends IOParameterValue
 final case class IOParameterValueBoolean(value: Boolean) extends IOParameterValue
 final case class IOParameterValueArray(array: Vector[IOParameterValue]) extends IOParameterValue
+final case class IOParameterValueHash(hash: Map[String, JsValue]) extends IOParameterValue
 final case class IOParameterValueDataObject(id: String,
                                             project: Option[String] = None,
                                             name: Option[String] = None,
@@ -142,6 +143,19 @@ object IOParameter {
   def parse(dxApi: DxApi, jsv: JsValue): IOParameter = {
     val fields = jsv.asJsObject.fields
 
+    // handle a job-based or workflow input reference
+    def createReference(fields: Map[String, JsValue]): DxIoParameterValueReference = {
+      fields(DxUtils.DxLinkKey) match {
+        case JsObject(innerFields) =>
+          DxIoParameterValueReference(innerFields.map {
+            case (key, JsString(value)) => key -> value
+            case other                  => throw new Exception(s"Invalid link field ${other}")
+          })
+        case other =>
+          throw new Exception(s"unexpected dx link value ${other}")
+      }
+    }
+
     def parseValue(key: String, value: JsValue, ioClass: DxIOClass.DxIOClass): IOParameterValue = {
       (ioClass, value) match {
         case (_, JsArray(array)) if key == DxIOSpec.Default && DxIOClass.isArray(ioClass) =>
@@ -154,17 +168,7 @@ object IOParameter {
             val dxObj = dxApi.dataObjectFromJson(link)
             IOParameterValue.forDataObjectField(key, Some(dxObj.id), dxObj.project.map(_.id))
           } catch {
-            case _: Throwable =>
-              // JBOR or some other reference
-              fields(DxUtils.DxLinkKey) match {
-                case JsObject(innerFields) =>
-                  DxIoParameterValueReference(innerFields.map {
-                    case (key, JsString(value)) => key -> value
-                    case other                  => throw new Exception(s"Invalid link field ${other}")
-                  })
-                case other =>
-                  throw new Exception(s"unexpected dx link value ${other}")
-              }
+            case _: Throwable => createReference(fields)
           }
         case (DxIOClass.File | DxIOClass.FileArray | DxIOClass.Other, JsObject(fields))
             if key != DxIOSpec.Default =>
@@ -184,6 +188,8 @@ object IOParameter {
         case (DxIOClass.File | DxIOClass.FileArray, JsString(s)) if key != DxIOSpec.Default =>
           val parsed = DxPath.parse(s)
           IOParameterValue.forDataObjectField(key, Some(parsed.name), parsed.projName)
+        case (_, JsObject(fields)) if fields.contains(DxUtils.DxLinkKey) =>
+          createReference(fields)
         case (DxIOClass.String | DxIOClass.StringArray, JsString(s)) =>
           IOParameterValueString(s)
         case (DxIOClass.Int | DxIOClass.IntArray, JsNumber(n)) if n.isValidLong =>
@@ -192,6 +198,8 @@ object IOParameter {
           IOParameterValueNumber(n)
         case (DxIOClass.Boolean | DxIOClass.BooleanArray, JsBoolean(b)) =>
           IOParameterValueBoolean(b)
+        case (DxIOClass.Hash | DxIOClass.HashArray, JsObject(fields)) =>
+          IOParameterValueHash(fields)
         case _ =>
           throw new Exception(s"Unexpected ${key} value ${jsv} of type ${ioClass}")
       }